+++ src/Header/Age.php
@@ -4,24 +4,45 @@ declare(strict_types = 1);
 namespace Innmind\Http\Header;
 
 use Innmind\Http\Header as HeaderInterface;
-use Innmind\Immutable\Sequence;
+use Innmind\Immutable\{
+    Sequence,
+    Maybe,
+};
 
 /**
  * @psalm-immutable
  */
 final class Age implements HeaderInterface
 {
-    public function __construct(
-        private AgeValue $value,
+    /**
+     * @param int<0, max> $age
+     */
+    private function __construct(
+        private int $age,
     ) {
     }
 
     /**
      * @psalm-pure
+     *
+     * @param int<0, max> $age
      */
     public static function of(int $age): self
     {
-        return new self(new AgeValue($age));
+        return new self($age);
+    }
+
+    /**
+     * @psalm-pure
+     *
+     * @return Maybe<self>
+     */
+    public static function maybe(int $age): Maybe
+    {
+        return Maybe::of(match (true) {
+            $age >= 0 => $age,
+            default => null,
+        })->map(static fn($age) => new self($age));
     }
 
     #[\Override]
@@ -37,11 +58,11 @@ final class Age implements HeaderInterface
     }
 
     /**
-     * @return 0|positive-int
+     * @return int<0, max>
      */
     public function age(): int
     {
-        return $this->value->age();
+        return $this->age;
     }
 
     #[\Override]
@@ -52,6 +73,6 @@ final class Age implements HeaderInterface
 
     private function header(): Header
     {
-        return new Header('Age', $this->value);
+        return new Header('Age', new Value\Value((string) $this->age));
     }
 }